All Locations as Objects
I have added all of the locations from the Heights in Canvas as objects in the gamemaker file. The next step is to add sprites to them all and place them all on maps. They are all organized in the objects file  the same way they are organized within the Heights.